Use lodash.debounce to improve magic line behaviour
Most logic is now executed on document ready (as before the previous commit).

There is a new reset helper function, which is sort of a debounced moveTo .caption.

The reset is executed on window load (to tackle to width changed by webfonts issue) and window resize.

Also reset is used as a hover leave event handler. It gives a nice side effect of little delay before the magic line goes back to initial position.

Bacause the delay, the reset has to be canceled by enter handler. Otherwise the magic line would randomly jump to initial position.