Display get in touch menu as block on mobile
It used to be a flex box on all form factors, but there were problems on older browsers. Besides that way it's easier to tap on a link (bigger touch target area).

The downside is that the page doesn't fit into most mobile screens. E.g. on iPhone 4 only first two items are visible.