Use components to modularize teacup
Components are pieces of markup that are not meant to be rendered as separate HTML, but are used as (possible reusable) fragments of pages. In this porject these includes freelance and typeform. Also landing has been moved to components.

Source files for components are in `components` directory. They are loaded by `teacup` task defined in `gulpfile.d/teacup`.